pyproject.toml
[project]
name = "expense-tracker"
version = "0.1.0"
description = "A simple expense tracker"
authors = [
    {name = "Yang Di",email = "yang.di2015@outlook.com"}
]
readme = "README.md"
requires-python = ">=3.13,<4"
dependencies = [
    "flask (>=3.1.2,<4.0.0)",
    "sqlalchemy (>=2.0.43,<3.0.0)",
    "flask-smorest (>=0.46.1,<0.47.0)",
    "flask-sqlalchemy (>=3.1.1,<4.0.0)",
    "marshmallow (>=4.0.1,<5.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "flask-migrate (>=4.1.0,<5.0.0)",
    "flask-bcrypt (>=1.0.1,<2.0.0)",
    "flask-jwt-extended (>=4.7.1,<5.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 79
exclude = '''
/(
    \.venv
  | build
  | dist
  | __pycache__
  | migrations
)/
'''

[tool.poetry.group.dev.dependencies]
pre-commit = "^4.3.0"
flake8 = "^7.3.0"
black = "^25.1.0"

